d/p/conf-disable-bluez5.patch: Disable BlueZ support by default
PipeWire is intended to be a suitable replacement for PulseAudio
eventually, but at the moment only the video side is ready for production
use, with the audio side considered experimental.

For ordinary ALSA audio devices, it is harmless to have audio support
available in PipeWire, because both PipeWire and PulseAudio will
close the audio device when not in use, so they can coexist (as long
as you do not try to use both at the same time). However, Bluetooth
audio devices need to be configured even when not in active use, so
they would fight.

Patch taken from Fedora.

Thanks: Wim Taymans