Fix pulse tunnel unload segfaulting pipewire-pulse
This happens when a zeroconf service advertise a pulseaudio protocol
tunnel when it does not service it. That is nothing is listening on the
advertised port.
Then the tunnel module is unload but since commit f6e25c239 module-pulse-tunnel: don't unload module from pulse thread
this makes pipewire-pulse segfault. See https://gitlab.freedesktop.org/pipewire/pipewire/-/issues/3199

This patch is already included in the upstream git repository in master.